Add support for custom python environments (#55)
* Add support for custom python environments

* Remove boost

* Add support for using custom Python backend stub

* Fix custom python backend stub support

* Do not ignore return value of chdir and cwd

* Review edits + improve the error message when there is not enough shared memory

* Review edits

+++ src/python.cc
@@ -36,7 +36,6 @@
 #include <cstdio>
 #include <cstring>
 #include <ctime>
-#include <filesystem>
 #include <functional>
 #include <memory>
 #include <numeric>
@@ -44,6 +43,7 @@
 #include <string>
 #include <thread>
 #include <vector>
+#include "pb_env.h"
 #include "pb_utils.h"
 #include "shm_manager.h"
 #include "triton/backend/backend_common.h"
@@ -66,17 +66,17 @@
     }                                                                  \
   } while (false)
 
-#define RESPOND_ALL_AND_RETURN_IF_EXCEPTION(RESPONSES, RESPONSES_COUNT, X)     \
-  do {                                                                         \
-    try {                                                                      \
-      (X);                                                                     \
-    }                                                                          \
-    catch (const PythonBackendException& exception) {                          \
-      TRITONSERVER_Error* raarie_err__ = TRITONSERVER_ErrorNew(                \
-          TRITONSERVER_ERROR_INTERNAL, exception.err_->error_message.c_str()); \
-      SendErrorForResponses(RESPONSES, RESPONSES_COUNT, raarie_err__);         \
-      return nullptr;                                                          \
-    }                                                                          \
+#define RESPOND_ALL_AND_RETURN_IF_EXCEPTION(RESPONSES, RESPONSES_COUNT, X) \
+  do {                                                                     \
+    try {                                                                  \
+      (X);                                                                 \
+    }                                                                      \
+    catch (const PythonBackendException& exception) {                      \
+      TRITONSERVER_Error* raarie_err__ = TRITONSERVER_ErrorNew(            \
+          TRITONSERVER_ERROR_INTERNAL, exception.what());                  \
+      SendErrorForResponses(RESPONSES, RESPONSES_COUNT, raarie_err__);     \
+      return nullptr;                                                      \
+    }                                                                      \
   } while (false)
 
 #define RESPOND_AND_RETURN_IF_ERROR(REQUEST, X)                         \
@@ -98,27 +98,27 @@
     }                                                                   \
   } while (false)
 
-#define RESPOND_AND_RETURN_IF_EXCEPTION(REQUEST, X)                            \
-  do {                                                                         \
-    try {                                                                      \
-      (X);                                                                     \
-    }                                                                          \
-    catch (const PythonBackendException& exception) {                          \
-      TRITONSERVER_Error* rarie_err__ = TRITONSERVER_ErrorNew(                 \
-          TRITONSERVER_ERROR_INTERNAL, exception.err_->error_message.c_str()); \
-      TRITONBACKEND_Response* rarie_response__ = nullptr;                      \
-      LOG_IF_ERROR(                                                            \
-          TRITONBACKEND_ResponseNew(&rarie_response__, REQUEST),               \
-          "failed to create response");                                        \
-      if (rarie_response__ != nullptr) {                                       \
-        LOG_IF_ERROR(                                                          \
-            TRITONBACKEND_ResponseSend(                                        \
-                rarie_response__, TRITONSERVER_RESPONSE_COMPLETE_FINAL,        \
-                rarie_err__),                                                  \
-            "failed to send error response");                                  \
-      }                                                                        \
-      return rarie_err__;                                                      \
-    }                                                                          \
+#define RESPOND_AND_RETURN_IF_EXCEPTION(REQUEST, X)                     \
+  do {                                                                  \
+    try {                                                               \
+      (X);                                                              \
+    }                                                                   \
+    catch (const PythonBackendException& exception) {                   \
+      TRITONSERVER_Error* rarie_err__ = TRITONSERVER_ErrorNew(          \
+          TRITONSERVER_ERROR_INTERNAL, exception.what());               \
+      TRITONBACKEND_Response* rarie_response__ = nullptr;               \
+      LOG_IF_ERROR(                                                     \
+          TRITONBACKEND_ResponseNew(&rarie_response__, REQUEST),        \
+          "failed to create response");                                 \
+      if (rarie_response__ != nullptr) {                                \
+        LOG_IF_ERROR(                                                   \
+            TRITONBACKEND_ResponseSend(                                 \
+                rarie_response__, TRITONSERVER_RESPONSE_COMPLETE_FINAL, \
+                rarie_err__),                                           \
+            "failed to send error response");                           \
+      }                                                                 \
+      return rarie_err__;                                               \
+    }                                                                   \
   } while (false)
 
 #define GUARDED_RESPOND_IF_ERROR(RESPONSES, IDX, X)                     \
@@ -145,8 +145,7 @@
       }                                                                 \
       catch (const PythonBackendException& pb_exception) {              \
         TRITONSERVER_Error* err__ = TRITONSERVER_ErrorNew(              \
-            TRITONSERVER_ERROR_INTERNAL,                                \
-            pb_exception.err_->error_message.c_str());                  \
+            TRITONSERVER_ERROR_INTERNAL, pb_exception.what());          \
         LOG_IF_ERROR(                                                   \
             TRITONBACKEND_ResponseSend(                                 \
                 (RESPONSES)[IDX], TRITONSERVER_RESPONSE_COMPLETE_FINAL, \
@@ -165,8 +164,7 @@
     }                                                          \
     catch (const PythonBackendException& pb_exception) {       \
       TRITONSERVER_Error* rarie_err__ = TRITONSERVER_ErrorNew( \
-          TRITONSERVER_ERROR_INTERNAL,                         \
-          pb_exception.err_->error_message.c_str());           \
+          TRITONSERVER_ERROR_INTERNAL, pb_exception.what());   \
       return rarie_err__;                                      \
     }                                                          \
   } while (false)
@@ -178,6 +176,7 @@ struct BackendState {
   int64_t shm_default_byte_size;
   int64_t shm_growth_byte_size;
   int64_t stub_timeout_seconds;
+  std::unique_ptr<EnvironmentManager> env_manager;
 };
 
 class ModelState : public BackendModel {
@@ -188,16 +187,20 @@ class ModelState : public BackendModel {
   // Get backend state
   BackendState* StateForBackend() { return backend_state_; }
 
+  // Get the Python execution environment
+  std::string PythonExecutionEnv() { return python_execution_env_; }
+
  private:
   ModelState(TRITONBACKEND_Model* triton_model);
   BackendState* backend_state_;
+  std::string python_execution_env_;
 };
 
 TRITONSERVER_Error*
 CreateTritonErrorFromException(const PythonBackendException& pb_exception)
 {
   return TRITONSERVER_ErrorNew(
-      TRITONSERVER_ERROR_INTERNAL, pb_exception.err_->error_message.c_str());
+      TRITONSERVER_ERROR_INTERNAL, pb_exception.what());
 }
 
 class ModelInstanceState : public BackendModelInstance {
@@ -245,7 +248,7 @@ class ModelInstanceState : public BackendModelInstance {
 
 ModelInstanceState::ModelInstanceState(
     ModelState* model_state, TRITONBACKEND_ModelInstance* triton_model_instance)
-    : BackendModelInstance(model_state, triton_model_instance),
+    : BackendModelInstance(model_state, triton_model_instance), child_pid_(0),
       initialized_(false)
 {
 }
@@ -742,8 +745,15 @@ ModelInstanceState::SetupChildProcess()
   int64_t shm_default_size =
       model_state->StateForBackend()->shm_default_byte_size;
 
-  shm_pool_ = std::make_unique<SharedMemory>(
-      shm_region_name, shm_default_size, shm_growth_size, true /* truncate */);
+  try {
+    shm_pool_ = std::make_unique<SharedMemory>(
+        shm_region_name, shm_default_size, shm_growth_size,
+        true /* truncate */);
+  }
+  catch (const PythonBackendException& pb_exception) {
+    return TRITONSERVER_ErrorNew(
+        TRITONSERVER_ERROR_INTERNAL, pb_exception.what());
+  }
 
   // Child Mutex and CV
   pthread_mutex_t* child_mutex;
@@ -798,6 +808,35 @@ ModelInstanceState::SetupChildProcess()
             .c_str());
   }
 
+  // Path to the extracted Python env
+  std::string python_execution_env = "";
+
+  // Path to environment activation script
+  std::string path_to_activate;
+  std::string path_to_libpython;
+
+  if (model_state->PythonExecutionEnv() != "") {
+    try {
+      python_execution_env =
+          model_state->StateForBackend()->env_manager->ExtractIfNotExtracted(
+              model_state->PythonExecutionEnv());
+    }
+    catch (PythonBackendException& pb_exception) {
+      return TRITONSERVER_ErrorNew(
+          TRITONSERVER_ERROR_INTERNAL, pb_exception.what());
+    }
+    path_to_activate = python_execution_env + "/bin/activate";
+    path_to_libpython = python_execution_env + "/lib";
+    if (python_execution_env.length() > 0 && !FileExists(path_to_activate)) {
+      return TRITONSERVER_ErrorNew(
+          TRITONSERVER_ERROR_INTERNAL,
+          (std::string("Path ") + path_to_activate +
+           " does not exist. The Python environment should contain an "
+           "'activate' script.")
+              .c_str());
+    }
+  }
+
   parent_pid_ = getpid();
   pthread_mutex_lock(parent_mutex_);
 
@@ -810,22 +849,48 @@ ModelInstanceState::SetupChildProcess()
 
   // Child process
   if (pid == 0) {
-    const char* stub_args[7];
-    stub_args[6] = nullptr;  // Last argument must be nullptr
+    const char* stub_args[4];
+    stub_args[0] = "bash";
+    stub_args[1] = "-c";
+    stub_args[3] = nullptr;  // Last argument must be nullptr
+
+    // Default Python backend stub
+    std::string python_backend_stub =
+        model_state->StateForBackend()->python_lib +
+        "/triton_python_backend_stub";
+
+    // Path to alternative Python backend stub
+    std::string model_python_backend_stub =
+        std::string(model_path) + "/triton_python_backend_stub";
+
+    if (FileExists(model_python_backend_stub)) {
+      python_backend_stub = model_python_backend_stub;
+    }
+
     std::stringstream ss;
-    ss << model_state->StateForBackend()->python_lib
-       << "/triton_python_backend_stub";
-    std::string stub_path = ss.str();
-    stub_args[0] = stub_path.c_str();
-    stub_args[1] = model_path_.c_str();
-    stub_args[2] = shm_region_name.c_str();
-    stub_args[3] = std::to_string(shm_default_size).c_str();
-    stub_args[4] = std::to_string(shm_growth_size).c_str();
-    stub_args[5] = std::to_string(parent_pid_).c_str();
-    if (execvp(stub_args[0], (char**)stub_args) == -1) {
+    ss << python_backend_stub << " " << model_path_ << " " << shm_region_name
+       << " " << shm_default_size << " " << shm_growth_size << " "
+       << parent_pid_ << " " << model_state->StateForBackend()->python_lib;
+
+    std::string bash_argument;
+    bash_argument = ss.str();
+    if (model_state->PythonExecutionEnv() != "") {
+      // Need to properly set the LD_LIBRARY_PATH so that Python environments
+      // using different python versions load properly.
+      bash_argument = "export LD_LIBRARY_PATH=" + path_to_libpython +
+                      ":$LD_LIBRARY_PATH; source " + path_to_activate + " && " +
+                      bash_argument;
+      LOG_MESSAGE(
+          TRITONSERVER_LOG_VERBOSE,
+          (std::string("Started Python backend stub: ") + bash_argument)
+              .c_str());
+    }
+
+    stub_args[2] = bash_argument.c_str();
+    if (execvp("bash", (char**)stub_args) == -1) {
       std::stringstream ss;
       ss << "Failed to run python backend stub. Errno = " << errno << '\n'
-         << "Python backend stub path: " << stub_path << '\n'
+         << "Python backend stub path: " << python_backend_stub << '\n'
          << "Shared Memory Region Name: " << shm_region_name << '\n'
          << "Shared Memory Default Byte Size: " << shm_default_size << '\n'
          << "Shared Memory Growth Byte Size: " << shm_growth_size << '\n';
@@ -851,7 +916,7 @@ ModelInstanceState::SetupChildProcess()
     if (pthread_cond_timedwait(parent_cond_, parent_mutex_, &ts) != 0) {
       return TRITONSERVER_ErrorNew(
           TRITONSERVER_ERROR_INTERNAL,
-          (std::string("Timed out occured while waiting for the stub process. "
+          (std::string("Timed out occurred while waiting for the stub process. "
                        "Failed to initialize model instance ") +
            Name())
               .c_str());
@@ -909,13 +974,15 @@ ModelInstanceState::~ModelInstanceState()
 
     // Wait for child notification
     pthread_cond_wait(parent_cond_, parent_mutex_);
+    pthread_mutex_unlock(parent_mutex_);
   }
 
-  // Terminate the child process.
-  int status;
-  kill(child_pid_, SIGTERM);
-  waitpid(child_pid_, &status, 0);
-  pthread_mutex_unlock(parent_mutex_);
+  // Terminate the child process if it has been created.
+  if (child_pid_ != 0) {
+    int status;
+    kill(child_pid_, SIGTERM);
+    waitpid(child_pid_, &status, 0);
+  }
 }
 
 TRITONSERVER_Error*
@@ -1003,10 +1070,23 @@ ModelState::ModelState(TRITONBACKEND_Model* triton_model)
   TRITONBACKEND_ArtifactType artifact_type;
   THROW_IF_BACKEND_MODEL_ERROR(
       TRITONBACKEND_ModelRepository(triton_model, &artifact_type, &path));
+  python_execution_env_ = "";
 
   void* bstate;
   THROW_IF_BACKEND_MODEL_ERROR(TRITONBACKEND_BackendState(backend, &bstate));
   backend_state_ = reinterpret_cast<BackendState*>(bstate);
+  triton::common::TritonJson::Value params;
+  if (model_config_.Find("parameters", &params)) {
+    // Skip the EXECUTION_ENV_PATH variable if it doesn't exist.
+    TRITONSERVER_Error* error =
+        GetParameterValue(params, "EXECUTION_ENV_PATH", &python_execution_env_);
+    if (error == nullptr) {
+      LOG_MESSAGE(
+          TRITONSERVER_LOG_INFO,
+          (std::string("Using Python execution env ") + python_execution_env_)
+              .c_str());
+    }
+  }
 
   if (artifact_type != TRITONBACKEND_ARTIFACT_FILESYSTEM) {
     throw triton::backend::BackendModelException(TRITONSERVER_ErrorNew(
@@ -1144,6 +1224,7 @@ TRITONBACKEND_Initialize(TRITONBACKEND_Backend* backend)
   RETURN_IF_ERROR(
       TRITONBACKEND_BackendArtifacts(backend, &artifact_type, &location));
   backend_state->python_lib = location;
+  backend_state->env_manager = std::make_unique<EnvironmentManager>();
 
   RETURN_IF_ERROR(TRITONBACKEND_BackendSetState(
       backend, reinterpret_cast<void*>(backend_state.get())));
